fetched the trailer of the movie using custom hook for background video of browse page
