+++ src/components/VideoBackground.js
@@ -1,25 +1,15 @@
-import React, { useEffect } from 'react'
-import { API_OPTIONS } from '../utils/constants'
+import React from 'react'
+import useGetTrailerVideo from '../hooks/useGetTrailerVideo'
+import { useSelector } from 'react-redux';
 
-const VideoBackground = ({ id }) => {
-    const getMovieTrailer = async () => {
-        const data = await fetch('https://api.themoviedb.org/3/movie/939243/videos?language=en-US', API_OPTIONS)
-        const json = await data.json();
-        console.log(json)
+const VideoBackground = ({ movieId }) => {
+    useGetTrailerVideo(movieId);
+    const trailerVideo = useSelector((store) => store.movies?.trailerVideo);
 
-        const filteredData = json.results.filter((video) => video.type === "Trailer");
-        const trailerData = filteredData.length ? filteredData[0] : json.results[0];
-        console.log(trailerData);
-
-
-
-    }
-    useEffect(() => {
-        getMovieTrailer();
-    }, [])
     return (
         <div>
             VideoBackground
+            <iframe width="560" height="315" src={"https://www.youtube.com/embed/" + trailerVideo?.key} title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerPolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
 
         </div>
     )
