+++ src/hooks/useNowPlayingMovies.js
@@ -10,7 +10,7 @@ const useNowPlayingMovies = () => {
     const nowPlayingMovies = async () => {
         const data = await fetch('https://api.themoviedb.org/3/movie/now_playing?page=1', API_OPTIONS);
         const json = await data.json();
-        console.log(json.results)
+        // console.log(json.results)
         dispatch(addNowPlayingMovies(json.results))
     }
