+++ src/utils/moviesSlice.js
@@ -3,14 +3,18 @@ const moviesSlice = createSlice({
     name: "movies",
     initialState: {
         nowPlayingMovies: null,
+        trailerVideo: null
     },
     reducers: {
         addNowPlayingMovies: (state, action) => {
             state.nowPlayingMovies = action.payload
+        },
+        addTrailerVideo: (state, action) => {
+            state.trailerVideo = action.payload
         }
     }
 
 })
 
-export const { addNowPlayingMovies } = moviesSlice.actions;
+export const { addNowPlayingMovies, addTrailerVideo } = moviesSlice.actions;
 export default moviesSlice.reducer;
